+++ sdk/storage/azblob/blob/models.go
@@ -216,16 +216,6 @@ func (o *downloadOptions) layoutAwareRoutingEnabled() bool {
 	return o.LayoutAwareRouting != LayoutAwareRoutingDisabled
 }
 
-func (o *downloadOptions) getBlobPropertiesOptions() *GetPropertiesOptions {
-	if o == nil {
-		return nil
-	}
-	return &GetPropertiesOptions{
-		AccessConditions: o.AccessConditions,
-		CPKInfo:          o.CPKInfo,
-	}
-}
-
 func (o *downloadOptions) getDownloadBlobOptions(rnge HTTPRange, rangeGetContentMD5 *bool) *DownloadStreamOptions {
 	if o == nil {
 		return nil